requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.2.0; python_version >= '3.8'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
boto==2.49.0
boto3==1.34.44; python_version >= '3.8'
botocore==1.34.44; python_version >= '3.8'
certifi==2024.2.2; python_version >= '3.6'
cffi==1.16.0; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cryptography==42.0.3; python_version >= '3.7'
fastapi==0.109.2; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
idna==3.6; python_version >= '3.5'
jinja2==3.1.3; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
moto==5.0.1; python_version >= '3.8'
mysql-connector-python==8.3.0; python_version >= '3.8'
pycparser==2.21
pydantic==2.6.1; python_version >= '3.8'
pydantic-core==2.16.2; python_version >= '3.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyyaml==6.0.1; python_version >= '3.6'
redis==5.0.1; python_version >= '3.7'
requests==2.31.0; python_version >= '3.7'
responses==0.25.0; python_version >= '3.8'
s3transfer==0.10.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
soupsieve==2.5; python_version >= '3.8'
starlette==0.36.3; python_version >= '3.8'
typing-extensions==4.9.0; python_version >= '3.8'
urllib3==2.0.7; python_version >= '3.7'
uvicorn==0.27.1; python_version >= '3.8'
werkzeug==3.0.1; python_version >= '3.8'
xmltodict==0.13.0; python_version >= '3.4'
